NX-163: remove plugin manager (step 1) - move registration of static resources to plugin module, use mediator to register repository types
