+++ components/nexus-core/src/main/java/org/sonatype/nexus/plugins/PluginDescriptor.java
@@ -17,13 +17,11 @@ import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
 
-import org.sonatype.nexus.proxy.registry.RepositoryTypeDescriptor;
-import org.sonatype.nexus.web.WebResource;
 import org.sonatype.plugin.metadata.GAVCoordinate;
 import org.sonatype.plugins.model.PluginMetadata;
 
 /**
- * Describes a Nexus plugin: its metadata, exports/imports, and what resources it contains.
+ * Describes a Nexus plugin: its metadata and exports/imports.
  */
 @Deprecated
 public final class PluginDescriptor
@@ -47,10 +45,6 @@ public final class PluginDescriptor
   private List<GAVCoordinate> importedPlugins = Collections.emptyList();
 
   private List<GAVCoordinate> resolvedPlugins = Collections.emptyList();
-
-  private List<RepositoryTypeDescriptor> repositoryTypes = Collections.emptyList();
-
-  private List<WebResource> staticResources = Collections.emptyList();
 
   // ----------------------------------------------------------------------
   // Constructors
@@ -96,26 +90,6 @@ public final class PluginDescriptor
 
     // TODO: list components? list exports/imports?
 
-    if (repositoryTypes != null) {
-      buf.append(LS);
-      buf.append("         Custom repository types:").append(LS);
-
-      for (final RepositoryTypeDescriptor type : repositoryTypes) {
-        buf.append("         * Repository type \"").append(type.getRole().getName());
-        buf.append("\", to be published at path \"").append(type.getPrefix()).append("\"").append(LS);
-      }
-    }
-
-    if (staticResources != null) {
-      buf.append(LS);
-      buf.append("         Static resources:").append(LS);
-
-      for (final WebResource resource : staticResources) {
-        buf.append("         * Content type \"").append(resource.getContentType());
-        buf.append("\", to be published at path \"").append(resource.getPath()).append("\"").append(LS);
-      }
-    }
-
     if (importedPlugins != null) {
       buf.append(LS);
       buf.append("         Imported plugins:").append(LS);
@@ -155,13 +129,5 @@ public final class PluginDescriptor
 
   void setResolvedPlugins(final List<GAVCoordinate> plugins) {
     resolvedPlugins = Collections.unmodifiableList(new ArrayList<GAVCoordinate>(plugins));
-  }
-
-  void setRepositoryTypes(final List<RepositoryTypeDescriptor> types) {
-    repositoryTypes = Collections.unmodifiableList(new ArrayList<RepositoryTypeDescriptor>(types));
-  }
-
-  void setStaticResources(final List<WebResource> resources) {
-    staticResources = Collections.unmodifiableList(new ArrayList<WebResource>(resources));
   }
 }
