+++ components/nexus-core/src/test/java/org/sonatype/nexus/plugins/DefaultNexusPluginManagerTest.java
@@ -19,11 +19,9 @@ import java.util.HashMap;
 import java.util.Map;
 
 import org.sonatype.nexus.guice.AbstractInterceptorModule;
-import org.sonatype.nexus.mime.MimeSupport;
 import org.sonatype.nexus.plugins.repository.NexusPluginRepository;
 import org.sonatype.nexus.plugins.repository.NoSuchPluginRepositoryArtifactException;
 import org.sonatype.nexus.plugins.repository.PluginRepositoryArtifact;
-import org.sonatype.nexus.proxy.registry.RepositoryTypeRegistry;
 import org.sonatype.plugin.metadata.GAVCoordinate;
 import org.sonatype.plugins.model.PluginDependency;
 import org.sonatype.plugins.model.PluginMetadata;
@@ -49,9 +47,6 @@ public class DefaultNexusPluginManagerTest
 {
 
   @Mock
-  private RepositoryTypeRegistry repositoryTypeRegistry;
-
-  @Mock
   private EventBus eventBus;
 
   @Mock
@@ -60,16 +55,13 @@ public class DefaultNexusPluginManagerTest
   @Mock
   private NexusPluginRepository nexusPluginRepository;
 
-  @Mock
-  private MimeSupport mimeSupport;
-
   @Test
   public void pluginDependenciesAreActivatedByGA()
       throws Exception
   {
     final DefaultNexusPluginManager underTest = new DefaultNexusPluginManager(
-        repositoryTypeRegistry, eventBus, pluginRepositoryManager,
-        new DefaultPlexusContainer(), mimeSupport, new HashMap<String, String>(),
+        eventBus, pluginRepositoryManager,
+        new DefaultPlexusContainer(), new HashMap<String, String>(),
         Collections.<AbstractInterceptorModule>emptyList()
     )
     {
