+++ testsupport/nexus-plugin-testsupport/src/main/java/org/sonatype/nexus/AbstractPluginTestCase.java
@@ -19,10 +19,11 @@ import java.util.Arrays;
 import java.util.HashMap;
 import java.util.List;
 
-import org.sonatype.nexus.guice.NexusAnnotatedBeanModule;
+import org.sonatype.nexus.guice.NexusTypeBinder;
 import org.sonatype.nexus.proxy.registry.RepositoryTypeDescriptor;
 
 import org.codehaus.plexus.DefaultPlexusContainer;
+import org.eclipse.sisu.plexus.PlexusAnnotatedBeanModule;
 import org.eclipse.sisu.plexus.PlexusBeanModule;
 import org.eclipse.sisu.space.ClassSpace;
 import org.eclipse.sisu.space.URLClassSpace;
@@ -53,12 +54,10 @@ public abstract class AbstractPluginTestCase
         scanList.add(getTestFile(sourceDir).toURI().toURL());
       }
 
-      final List<RepositoryTypeDescriptor> repositoryTypes = new ArrayList<RepositoryTypeDescriptor>();
-
       final ClassSpace annSpace =
           new URLClassSpace(getContainer().getContainerRealm(), scanList.toArray(new URL[scanList.size()]));
-      final NexusAnnotatedBeanModule nexusPluginModule =
-          new NexusAnnotatedBeanModule(annSpace, new HashMap<String, String>(), repositoryTypes);
+      final PlexusBeanModule nexusPluginModule =
+          new PlexusAnnotatedBeanModule(annSpace, new HashMap<String, String>()).with(NexusTypeBinder.STRATEGY);
       final List<PlexusBeanModule> modules = Arrays.<PlexusBeanModule>asList(nexusPluginModule);
 
       // register new injector
